Make it run in a browser, add a makefile, clean up a bit
