+++ public/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset=utf-8>
+    <title>Prototypal Inheritance</title>
+    <link rel="stylesheet" type="text/css" href="css/base.css">
+    <script src="js/bundle.js"></script>
+  </head>
+  <body>
+    <div id="game-canvas">
+    </div>
+
+    <script type="text/javascript">
+      // Game.init();
+    </script>
+  </body>
+</html>
